spssformattype.h
Nominal values imported as string if column contains value with a non
zero fractional part (i.e. is not integer value).

+++ JASP-Desktop/importers/spss/spssrecinter.h
@@ -82,7 +82,7 @@ public:
 
 	READ_ATTR(std::string, spssName)	// The name as in hte file.
 
-	READ_ATTR(size_t, spssStringLen)		// Length of the string (if string).
+	READ_ATTR(long, spssStringLen)		// Length of the string (if string).
 
 	READ_ATTR(size_t, columnSpan)			// Number of data cells this column spans.
 	void incrementColumnSpan() { _columnSpan++; }
